store/wizard: extract details types to submodule
Same pattern as the filesystem extraction — types go first to
establish the submodule boundary.
