+++ src/store/slices/wizard/details/index.ts
@@ -1,0 +1,1 @@
+export * from './types';
